[540431] getToStringBufferSize() return size_t instead of unsigned int
Change unsigned int type for size_t in all ocurrences of
getToStringBufferSize()

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=540431

+++ src/core/datatypes/forte_any.cpp
@@ -239,6 +239,6 @@ const TForteByte CIEC_ANY::csmStringBufferSize[] = {
          0 /*e_Max*/
     };
 
-unsigned int CIEC_ANY::getToStringBufferSize() const{
+size_t CIEC_ANY::getToStringBufferSize() const {
   return csmStringBufferSize[getDataTypeID()];
 }
